LayoutTests/imported/w3c/web-platform-tests/trusted-types/empty-default-policy-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.


Harness Error (TIMEOUT), message = null

TIMEOUT Count SecurityPolicyViolation events. Test timed out
FAIL script.src default assert_throws_js: function "_ => element[c[1]] = "abc"" did not throw
FAIL div.innerHTML default assert_throws_js: function "_ => element[c[1]] = "abc"" did not throw
FAIL script.text default assert_throws_js: function "_ => element[c[1]] = "abc"" did not throw
